requirements.txt
Flask==3.1.1
gunicorn==21.2.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
python-dotenv==1.0.1
